+++ src/main/java/com/ytdd9527/networksexpansion/core/items/machines/AbstractManualCrafter.java
@@ -5,6 +5,7 @@ import com.balugaq.netex.api.enums.FeedbackType;
 import com.balugaq.netex.utils.BlockMenuUtil;
 import com.ytdd9527.networksexpansion.core.items.SpecialSlimefunItem;
 import com.ytdd9527.networksexpansion.utils.itemstacks.ItemStackUtil;
+import io.github.sefiraat.networks.Networks;
 import io.github.sefiraat.networks.events.NetworkCraftEvent;
 import io.github.sefiraat.networks.slimefun.network.AdminDebuggable;
 import io.github.sefiraat.networks.utils.StackUtils;
@@ -169,7 +170,7 @@ public abstract class AbstractManualCrafter extends SpecialSlimefunItem implemen
                 SlimefunItem sfi = SlimefunItem.getByItem(item);
                 if (sfi != null) {
                     if (sfi.isDisabled() || sfi.isDisabledIn(world)) {
-                        player.sendMessage(ChatColor.RED + "输出物已被禁用"); // todo: add to lang
+                        player.sendMessage(Networks.getLocalizationService().getString("messages.unsupported-operation.manual_crafter.disabled-output"));
                         sendFeedback(blockMenu.getLocation(), FeedbackType.DISABLED_OUTPUT);
                         continue;
                     }
@@ -184,7 +185,7 @@ public abstract class AbstractManualCrafter extends SpecialSlimefunItem implemen
 
                 ItemStack left = BlockMenuUtil.pushItem(blockMenu, ItemStackUtil.getCleanItem(item), getOutputSlots());
                 if (left != null && left.getType() != Material.AIR) {
-                    player.sendMessage(ChatColor.RED + "输出槽已满"); // todo: add to lang
+                    player.sendMessage(Networks.getLocalizationService().getString("messages.unsupported-operation.manual_crafter.full-output"));
                     sendFeedback(blockMenu.getLocation(), FeedbackType.NO_ENOUGH_SPACE);
                     world.dropItem(blockMenu.getLocation(), left);
                 }
@@ -194,7 +195,7 @@ public abstract class AbstractManualCrafter extends SpecialSlimefunItem implemen
         if (recipe.getConsumeEnergy() > 0) {
             removeCharge(blockMenu.getLocation(), recipe.getConsumeEnergy());
         }
-        player.sendMessage(ChatColor.GREEN + "合成成功"); // todo: add to lang
+        player.sendMessage(Networks.getLocalizationService().getString("messages.completed-operation.manual_crafter.success"));
         sendFeedback(blockMenu.getLocation(), FeedbackType.SUCCESS);
         return true;
     }
@@ -274,14 +275,14 @@ public abstract class AbstractManualCrafter extends SpecialSlimefunItem implemen
                 SlimefunItem sfi = SlimefunItem.getByItem(item);
                 if (sfi != null) {
                     if (sfi.isDisabled() || sfi.isDisabledIn(world)) {
-                        player.sendMessage(ChatColor.RED + "输出物已被禁用"); // todo: add to lang
+                        player.sendMessage(Networks.getLocalizationService().getString("messages.unsupported-operation.manual_crafter.disabled-output"));
                         sendFeedback(blockMenu.getLocation(), FeedbackType.DISABLED_OUTPUT);
                         continue;
                     }
                 }
                 ItemStack left = BlockMenuUtil.pushItem(blockMenu, item, getOutputSlots());
                 if (left != null && left.getType() != Material.AIR) {
-                    player.sendMessage(ChatColor.RED + "输出槽已满"); // todo: add to lang
+                    player.sendMessage(Networks.getLocalizationService().getString("messages.unsupported-operation.manual_crafter.full-output"));
                     sendFeedback(blockMenu.getLocation(), FeedbackType.NO_ENOUGH_SPACE);
                     world.dropItem(blockMenu.getLocation(), left);
                 }
@@ -291,7 +292,7 @@ public abstract class AbstractManualCrafter extends SpecialSlimefunItem implemen
         if (recipe.getConsumeEnergy() > 0) {
             removeCharge(blockMenu.getLocation(), recipe.getConsumeEnergy());
         }
-        player.sendMessage(ChatColor.GREEN + "合成成功"); // todo: add to lang
+        player.sendMessage(Networks.getLocalizationService().getString("messages.completed-operation.manual_crafter.success"));
         sendFeedback(blockMenu.getLocation(), FeedbackType.SUCCESS);
         return true;
     }
